pyproject.toml
[project]
name = "freeway-dqn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.8.0",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.12.5",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
    "torchvision>=0.24.1",
]
